hw 6.3 done and 6.4 v2

+++ 06_InheritanceAndPolymorphism/Enterprise/src/Employee.java
@@ -2,6 +2,6 @@ import java.math.BigDecimal;
 
 public interface Employee {
 
-    public BigDecimal getMonthSalary();
+    public int getMonthSalary();
 
 }
