+++ 06_InheritanceAndPolymorphism/Enterprise/src/Operator.java
@@ -1,16 +1,15 @@
-import java.math.BigDecimal;
 
 public class Operator implements Employee{
 
     private Company company;
-    private BigDecimal salary = BigDecimal.valueOf(42000);
+    private int salary = 42000;
 
     public void setCompany(Company company) {
         this.company = company;
     }
 
     @Override
-    public BigDecimal getMonthSalary() {
+    public int getMonthSalary() {
         return salary;
     }
 }
